unified_trading_system/requirements.txt
# Core dependencies (from working installation)
aiohttp==3.13.5
numpy==2.4.4
pandas==3.0.2
PyYAML==6.0.1
python-dotenv==1.2.2

# Observability
prometheus_client==0.25.0
psutil>=5.9.0

# Machine Learning
torch==2.11.0+cpu
xgboost==3.2.0
scikit-learn>=1.5.0
scipy==1.17.1

# Data Feeds
websockets==10.4.0
requests==2.31.0

# Optional Performance
uvloop>=0.21.0